+++ crates/ruma-common/CHANGELOG.md
@@ -29,10 +29,12 @@ Improvements:
 - `UserProfile::merge` has been renamed to `UserProfile::apply`.
 - Add experimental support for [MSC4495] (Selective Presence).
 - Add experimental support for [MSC4494] (Membership-based invite blocking).
+- Add unstable support for [MSC4527] (State event power levels).
 
 [MSC4438]: https://github.com/matrix-org/matrix-spec-proposals/pull/4438
 [MSC4495]: https://github.com/matrix-org/matrix-spec-proposals/pull/4495
 [MSC4494]: https://github.com/matrix-org/matrix-spec-proposals/pull/4494
+[MSC4527]: https://github.com/matrix-org/matrix-spec-proposals/pull/4527
 
 ## 0.19.0